refactor(core): migrate GitHub logic to core.api
- Introduce core.api with API client, Release/Asset models, and
  high level GitHub client
- Update imports across the codebase to use core.api instead of
  core.github
- Add platform compatibility constants:
  INCOMPATIBLE_PLATFORM_PATTERNS and INCOMPATIBLE_PLATFORM_EXTENSIONS
- Remove obsolete core/github modules
