--- src/my_unicorn/core/github/models.py
@@ -1,268 +1,0 @@
-"""GitHub asset and release models."""
-
-from __future__ import annotations
-
-import re
-from dataclasses import dataclass, replace
-from typing import Any
-
-from my_unicorn.utils.asset_validation import is_appimage_file
-
-
-def extract_and_validate_version(package_string: str) -> str | None:
-    """Extract and validate version from package string.
-
-    Combines extraction, sanitization, and validation in one function.
-    Used by GitHub models to normalize version strings from release data.
-
-    Args:
-        package_string: Package string that may contain version
-
-    Returns:
-        Valid version string or None if extraction/validation fails
-
-    """
-    if not package_string:
-        return None
-
-    # Handle package@version format
-    if "@" in package_string:
-        parts = package_string.split("@")
-        if len(parts) < 2:
-            return None
-        version = parts[-1].strip()
-        if not version:
-            return None
-    else:
-        version = package_string
-
-    # Sanitize version
-    version = version.lstrip("v").replace("@", "").strip("\"'").strip()
-    if not version:
-        return None
-
-    # Validate version
-    pattern = r"^\d+(\.\d+)*(-[a-zA-Z0-9.-]+)?$"
-    if re.match(pattern, version):
-        return version
-    return None
-
-
-@dataclass(slots=True, frozen=True)
-class Asset:
-    """Represents a GitHub release asset.
-
-    Attributes:
-        name: Asset filename
-        size: Asset size in bytes
-        digest: Asset digest/hash (may be empty)
-        browser_download_url: Direct download URL for the asset
-
-    """
-
-    name: str
-    size: int
-    digest: str
-    browser_download_url: str
-
-    @classmethod
-    def from_api_response(cls, asset_data: dict[str, Any]) -> Asset | None:
-        """Create Asset from GitHub API response data.
-
-        Args:
-            asset_data: Raw asset data from GitHub API
-
-        Returns:
-            Asset instance or None if required fields are missing
-
-        """
-        try:
-            name = asset_data.get("name", "")
-            size = asset_data.get("size", 0)
-            download_url = asset_data.get("browser_download_url", "")
-            digest = asset_data.get("digest", "")
-
-            if not name or not download_url:
-                return None
-
-            return cls(
-                name=name,
-                size=int(size),
-                digest=digest,
-                browser_download_url=download_url,
-            )
-        except (KeyError, TypeError, ValueError):
-            return None
-
-    def is_appimage(self) -> bool:
-        """Check if this asset is an AppImage file.
-
-        Returns:
-            True if asset is an AppImage, False otherwise
-
-        """
-        return is_appimage_file(self.name)
-
-
-@dataclass(slots=True, frozen=True)
-class Release:
-    """Represents a GitHub release with its metadata and assets.
-
-    Attributes:
-        owner: Repository owner
-        repo: Repository name
-        version: Normalized version string
-        prerelease: Whether this is a prerelease
-        assets: List of release assets
-        original_tag_name: Original tag name from GitHub
-
-    """
-
-    owner: str
-    repo: str
-    version: str
-    prerelease: bool
-    assets: list[Asset]
-    original_tag_name: str
-
-    @classmethod
-    def from_api_response(
-        cls, owner: str, repo: str, api_data: dict[str, Any]
-    ) -> Release:
-        """Create Release from GitHub API response data.
-
-        Args:
-            owner: Repository owner
-            repo: Repository name
-            api_data: Raw release data from GitHub API
-
-        Returns:
-            Release instance
-
-        """
-        from my_unicorn.core.github.models import Asset  # noqa: PLC0415
-
-        tag_name = api_data.get("tag_name", "")
-        version = cls._normalize_version(tag_name)
-        prerelease = api_data.get("prerelease", False)
-
-        # Convert assets
-        assets = []
-        for asset_data in api_data.get("assets", []):
-            asset = Asset.from_api_response(asset_data)
-            if asset:
-                assets.append(asset)
-
-        return cls(
-            owner=owner,
-            repo=repo,
-            version=version,
-            prerelease=prerelease,
-            assets=assets,
-            original_tag_name=tag_name,
-        )
-
-    @staticmethod
-    def _normalize_version(tag_name: str) -> str:
-        """Normalize version by extracting and sanitizing version string.
-
-        Args:
-            tag_name: Version tag that may have 'v' prefix or package format
-
-        Returns:
-            Sanitized version string
-
-        """
-        if not tag_name:
-            return ""
-
-        normalized = extract_and_validate_version(tag_name)
-        if normalized is None:
-            return tag_name.lstrip("v")
-
-        return normalized
-
-    def to_dict(self) -> dict[str, Any]:
-        """Convert Release to dictionary for caching.
-
-        Returns:
-            Dictionary representation suitable for JSON serialization
-
-        """
-        return {
-            "owner": self.owner,
-            "repo": self.repo,
-            "version": self.version,
-            "prerelease": self.prerelease,
-            "assets": [
-                {
-                    "name": asset.name,
-                    "size": asset.size,
-                    "digest": asset.digest,
-                    "browser_download_url": asset.browser_download_url,
-                }
-                for asset in self.assets
-            ],
-            "original_tag_name": self.original_tag_name,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> Release:
-        """Create Release from dictionary (e.g., from cache).
-
-        Args:
-            data: Dictionary representation of a release
-
-        Returns:
-            Release instance
-
-        """
-        from my_unicorn.core.github.models import Asset  # noqa: PLC0415
-
-        assets = [
-            Asset(
-                name=a["name"],
-                size=a["size"],
-                digest=a.get("digest", ""),
-                browser_download_url=a["browser_download_url"],
-            )
-            for a in data.get("assets", [])
-        ]
-
-        return cls(
-            owner=data["owner"],
-            repo=data["repo"],
-            version=data["version"],
-            prerelease=data["prerelease"],
-            assets=assets,
-            original_tag_name=data["original_tag_name"],
-        )
-
-    def get_appimages(self) -> list[Asset]:
-        """Get all AppImage assets from this release.
-
-        Returns:
-            List of AppImage assets
-
-        """
-        return [asset for asset in self.assets if asset.is_appimage()]
-
-    def filter_for_platform(self) -> Release:
-        """Return new Release with only platform-relevant assets.
-
-        Filters assets to include only:
-        - Linux x86_64 AppImages (excludes Windows, macOS, ARM)
-        - Checksum files for compatible AppImages
-
-        This method uses AssetSelector.filter_for_cache() for consistent
-        filtering across the application.
-
-        Returns:
-            New Release instance with filtered assets
-
-        """
-        # Import here to avoid circular imports
-        from my_unicorn.core.github import AssetSelector  # noqa: PLC0415
-
-        filtered_assets = AssetSelector.filter_for_cache(self.assets)
-        return replace(self, assets=filtered_assets)
