+++ src/my_unicorn/core/services/install_service.py
@@ -11,9 +11,9 @@ from typing import Any
 import aiohttp
 
 from my_unicorn.config.config import ConfigManager
+from my_unicorn.core.api import GitHubClient
 from my_unicorn.core.download import DownloadService
 from my_unicorn.core.file_ops import FileOperations
-from my_unicorn.core.github import GitHubClient
 from my_unicorn.core.install import InstallHandler
 from my_unicorn.core.post_download import PostDownloadProcessor
 from my_unicorn.core.protocols.progress import (
