+++ src/my_unicorn/core/update/context.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from my_unicorn.constants import ERROR_CATALOG_MISSING
-from my_unicorn.core.github import Asset, get_github_config
+from my_unicorn.core.api import Asset, get_github_config
 from my_unicorn.exceptions import ConfigurationError, UpdateError
 from my_unicorn.logger import get_logger
 from my_unicorn.utils.appimage_utils import select_best_appimage_asset
